migrate component
